requirements.txt
anyio==4.2.0
aiokafka==0.10.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-restful==0.5.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
psutil==5.9.8
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.1
python-dateutil==2.8.2
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.0
pyjwt==2.8.0
alembic==1.13.1
psycopg2-binary==2.9.9
rocketchat-API==1.32.0
pandas==2.2.1
redis==5.0.3
python-multipart==0.0.9
boto3==1.34.91
PILLOW
openpyxl
../nai_util